+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Development
 
 - Make parameters TEMPERATURE_AIR_MAX_200 and TEMPERATURE_AIR_MIN_200 summarizable/interpolatable
 - Add streamlit app for DWD climate stations
+- Add sql query function to streamlit app
 
 0.60.0 (16.09.2023)
 *******************